moved sgf.serialize from sgf/parser.go to sgf/serialize.go
